Split CachedSandboxState and extract tzkt part from it
