+++ taqueria-vscode-extension/src/lib/gui/CachedSandboxState.ts
@@ -1,50 +1,33 @@
-import { Protocol } from '@taqueria/node-sdk/types';
 import * as rpc from '@taquito/rpc';
 import { TezosToolkit } from '@taquito/taquito';
 import * as rxjs from 'rxjs';
-import * as signalR from '../../../signalr';
-import { VsCodeHelper } from '../helpers';
-import { OutputLevels } from '../LogHelper';
-import { getSandboxAccounts } from './helpers/SandboxDataHelpers';
-import { TzKTAccountData } from './helpers/TzKTFetcher';
 import { ObservableConfig } from './ObservableConfig';
-import { SandboxModel, TzKtHead } from './SandboxDataModels';
-
-const { Url } = Protocol;
+import { SandboxModel } from './SandboxDataModels';
 
 export type SandboxState = 'running' | 'stopped' | 'unknown';
 
 export class CachedSandboxState {
-	connection: signalR.HubConnection | undefined;
-	sandboxHead = new rxjs.BehaviorSubject<rpc.BlockHeaderResponse | undefined>(undefined);
-	indexerHead = new rxjs.BehaviorSubject<TzKtHead | undefined>(undefined);
-	taquito: TezosToolkit | undefined;
 	private _currentSandboxBaseUrl: string | undefined;
-	private _currentTzKtBaseAddress: string | undefined;
 	private _state: SandboxState;
+
+	sandBoxModel: SandboxModel;
+	sandboxHead = new rxjs.BehaviorSubject<rpc.BlockHeaderResponse | undefined>(undefined);
+	taquito: TezosToolkit | undefined;
 
 	async setState(value: SandboxState) {
 		if (this._state === value) {
 			return;
 		}
 		this._state = value;
-		if (value === 'running') {
-			await this.startConnection();
-		} else {
-			this.connection?.stop();
-		}
 	}
 
 	get state() {
 		return this._state;
 	}
 
-	public sandBoxModel: SandboxModel;
-
 	constructor(
-		private readonly helper: VsCodeHelper,
 		private readonly sandboxName: string,
-		private readonly containerName: string,
+		containerName: string,
 		private readonly observableConfig: ObservableConfig,
 		state: SandboxState,
 	) {
@@ -60,40 +43,8 @@ export class CachedSandboxState {
 		};
 		this._state = state;
 	}
-	async updateConfig(): Promise<void> {
-		this.updateSandboxBaseUrl();
-		await this.updateTzKtBaseUrl();
-	}
-	async updateTzKtBaseUrl() {
-		const tzKtBaseAddress = this.getTzKtBaseUrl(this.sandboxName);
-		if (this._currentTzKtBaseAddress === tzKtBaseAddress) {
-			return;
-		}
-		this.helper.logHelper.showLog(
-			OutputLevels.debug,
-			`TzKt address changed from ${this._currentTzKtBaseAddress} to ${tzKtBaseAddress}`,
-		);
-		this._currentTzKtBaseAddress = tzKtBaseAddress;
-		if (this.connection) {
-			await this.connection.stop();
-		}
-		this.connection = new signalR.HubConnectionBuilder()
-			.withUrl(`${tzKtBaseAddress}/v1/events`, {
-				skipNegotiation: true,
-				transport: signalR.HttpTransportType.WebSockets,
-			})
-			.configureLogging(signalR.LogLevel.Information)
-			.build();
 
-		this.connection.onclose(async () => {
-			await this.startConnection();
-		});
-
-		this.connection.on('head', data => this.onHeadFromTzKt(data));
-		this.connection.on('accounts', payload => this.onAccountsFromTzKt(payload));
-	}
-
-	updateSandboxBaseUrl(): void {
+	private updateSandboxBaseUrl(): void {
 		const sandboxBaseUrl = this.getSandboxBaseUrl(this.sandboxName);
 		if (sandboxBaseUrl === this._currentSandboxBaseUrl) {
 			return;
@@ -106,43 +57,8 @@ export class CachedSandboxState {
 		this.taquito = new TezosToolkit(sandboxBaseUrl);
 	}
 
-	private onHeadFromTzKt(data: { type: number; data: TzKtHead }): void {
-		this.helper.logHelper.showLog(OutputLevels.debug, JSON.stringify(data));
-		if (data.type === 1) {
-			this.headFromTzKt.next(data.data);
-		}
-	}
-
-	private onAccountsFromTzKt(data: { type: number; data: TzKTAccountData }): void {
-		this.helper.logHelper.showLog(OutputLevels.debug, JSON.stringify(data));
-		if (data.type === 1) {
-			this.accountsFromTzKt.next(data.data);
-		}
-	}
-
-	headFromTzKt = new rxjs.BehaviorSubject<TzKtHead | undefined>(undefined);
-	accountsFromTzKt = new rxjs.BehaviorSubject<TzKTAccountData | undefined>(undefined);
-
-	async startConnection() {
-		if (!this.connection) {
-			return;
-		}
-		try {
-			await this.connection.start();
-			await this.connection.invoke('SubscribeToHead');
-			// TODO: make sure to reconnect with new accounts if config.json changes
-			const sandboxAccounts = getSandboxAccounts(this.observableConfig.currentConfig, this.sandboxName);
-			await this.connection.invoke('SubscribeToAccounts', {
-				addresses: sandboxAccounts.map(account => account.address),
-			});
-			this.helper.logHelper.showLog(OutputLevels.debug, 'SignalR Connected');
-		} catch (err) {
-			this.helper.logHelper.showLog(OutputLevels.debug, 'Error while connecting SignalR');
-			this.helper.logAllNestedErrors(err, true);
-			setTimeout(() => {
-				this.startConnection();
-			}, 5000);
-		}
+	async updateConfig(): Promise<void> {
+		this.updateSandboxBaseUrl();
 	}
 
 	getSandboxBaseUrl(sandboxName: string): string | undefined {
@@ -154,15 +70,6 @@ export class CachedSandboxState {
 		if (!port) {
 			port = '80';
 		}
-		return `http://127.0.0.1:${port}`;
-	}
-
-	getTzKtBaseUrl(sandboxName: string): string | undefined {
-		const sandbox = this.observableConfig.currentConfig.config?.config.sandbox?.[sandboxName];
-		if (!sandbox || typeof sandbox === 'string') {
-			return undefined;
-		}
-		let port = sandbox.tzkt?.apiPort ?? 5000;
 		return `http://127.0.0.1:${port}`;
 	}
 }
